fix: aktiver fritekstbrev for saksbehandler på ferdigstilte saker
- Flytt hjemmel-kodeverk-endepunkt til SakKodeverkController der det hører hjemme,
  slik at GET /api/sak/kodeverk/hjemler fungerer korrekt
- Tillat oppretting av fritekstbrev når sak har SEND_KLAGE-rettighet (i tillegg til SAKSBEHANDLE),
  slik at saksbehandler kan sende fritekstbrev på ferdigstilte saker
- Oppdater frontend: fritekstbrev-dialogen sjekker nå SAKSBEHANDLE eller SEND_KLAGE
  for skriverettighet og aktiverer menyvalget tilsvarende
- Legg til suksesslogg i KlageService etter vellykket sending til Kabal
- Fiks Kabal-mock til å returnere korrekte felter (behandlingId, mottattDato)
