+++ backend/src/main/kotlin/no/nav/historisk/superhelt/sak/rest/SakKodeverkController.kt
@@ -1,9 +1,11 @@
 package no.nav.historisk.superhelt.sak.rest
 
 import io.swagger.v3.oas.annotations.Operation
+import no.nav.historisk.superhelt.klage.rest.HjemmelDto
 import no.nav.historisk.superhelt.sak.SakStatus
 import no.nav.historisk.superhelt.sak.StonadsType
 import no.nav.historisk.superhelt.vedtak.VedtaksResultat
+import no.nav.kabal.model.Hjemmel
 import org.springframework.web.bind.annotation.GetMapping
 import org.springframework.web.bind.annotation.RequestMapping
 import org.springframework.web.bind.annotation.RestController
@@ -11,6 +13,20 @@ import org.springframework.web.bind.annotation.RestController
 @RestController
 @RequestMapping("/api/sak/kodeverk")
 class SakKodeverkController {
+
+    @Operation(operationId = "getKodeverkHjemler")
+    @GetMapping("hjemler")
+    fun hjemlerKodeverk(): List<HjemmelDto> {
+        return Hjemmel.entries.map { hjemmel ->
+            HjemmelDto(
+                id = hjemmel.id,
+                lovKildeNavn = hjemmel.lovKilde.navn,
+                lovKildeBeskrivelse = hjemmel.lovKilde.beskrivelse,
+                spesifikasjon = hjemmel.spesifikasjon,
+                visningsnavn = "${hjemmel.lovKilde.beskrivelse} ${hjemmel.spesifikasjon}",
+            )
+        }
+    }
 
     @Operation(operationId = "getKodeverkStonadType",)
     @GetMapping("stonadtyper")
